now has math distance based movement for time it will take

+++ src/ProjectNeon/Assets/Core/UI/TweenedUIMovement/GoToTweenRequested.cs
@@ -4,9 +4,9 @@ public class GoToTweenRequested
 {
     public Transform Transform { get; }
     public Vector3 Destination { get; }
-    public int Seconds { get; }
+    public float Seconds { get; }
 
-    public GoToTweenRequested(Transform transform, Vector3 destination, int seconds)
+    public GoToTweenRequested(Transform transform, Vector3 destination, float seconds)
     {
         Transform = transform;
         Destination = destination;
